[reorg] Move MultipleSubst into new layout.
https://github.com/harfbuzz/harfbuzz/commit/dea0681db
