+++ src/hb/ot/layout/GSUB/mod.rs
@@ -1,1 +1,3 @@
+mod multi_subst;
+mod sequence;
 mod single_subst;
