+++ src/hb/ot/layout/GSUB/multi_subst.rs
@@ -1,0 +1,20 @@
+use crate::hb::ot_layout_gsubgpos::OT::hb_ot_apply_context_t;
+use crate::hb::ot_layout_gsubgpos::{Apply, WouldApply, WouldApplyContext};
+use ttf_parser::gsub::MultipleSubstitution;
+
+// MultipleSubstFormat1::would_apply
+impl WouldApply for MultipleSubstitution<'_> {
+    fn would_apply(&self, ctx: &WouldApplyContext) -> bool {
+        ctx.glyphs.len() == 1 && self.coverage.get(ctx.glyphs[0]).is_some()
+    }
+}
+
+// MultipleSubstFormat1::apply
+impl Apply for MultipleSubstitution<'_> {
+    fn apply(&self, ctx: &mut hb_ot_apply_context_t) -> Option<()> {
+        let glyph = ctx.buffer.cur(0).as_glyph();
+        let index = self.coverage.get(glyph)?;
+        let seq = self.sequences.get(index)?;
+        seq.apply(ctx)
+    }
+}
